Create activity resource
