+++ src/subapps/projects/components/Activities/ActivityCard.less
@@ -1,4 +1,4 @@
-@import '../../../shared/lib.less';
+@import '../../../../shared/lib.less';
 
 .activity-card {
   margin: 1em;
