+++ src/subapps/projects/components/Activities/ActivityForm.less
@@ -1,4 +1,4 @@
-@import '../../../shared/lib.less';
+@import '../../../../shared/lib.less';
 
 .activity-form {
   .ant-form-item-label {
